New activity ChoosePlatformActivity commit
