+++ app/src/main/AndroidManifest.xml
@@ -15,6 +15,8 @@
             android:theme="@style/AppTheme"
             android:usesCleartextTraffic="true"
             tools:ignore="AllowBackup,GoogleAppIndexingWarning,UnusedAttribute">
+        <activity android:name=".activities.ChoosePlatformActivity">
+        </activity>
         <activity
                 android:name=".activities.MainActivity"
                 android:label="@string/title_activity_main"
