+++ app/src/main/java/com/spanishinquisition/treecompany/activities/SplashScreenActivity.kt
@@ -8,25 +8,18 @@ import android.os.Bundle
 import android.util.Log
 import android.widget.ImageView
 import android.widget.TextView
-import android.widget.Toast
 import com.spanishinquisition.treecompany.R
-import com.spanishinquisition.treecompany.models.Platform
 import com.spanishinquisition.treecompany.rest.SplashConnection
-import com.spanishinquisition.treecompany.rest.getClient
 import io.reactivex.Observable
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.disposables.Disposable
 import io.reactivex.schedulers.Schedulers
 import kotlin.random.Random
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
 
 class SplashScreenActivity : Activity() {
     private lateinit var loadingTv: TextView
     private lateinit var imageHome: ImageView
     private lateinit var localhostSubscription: Disposable
-    private lateinit var platformItems: Array<Platform>
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -73,34 +66,16 @@ class SplashScreenActivity : Activity() {
 
     private fun verifyContact(isConnected: Boolean) {
         if (isConnected) {
-            getPlatforms()
-            getPrefPlatform()
-            startActivity(Intent(this, MainActivity::class.java))
+            val firstTime = getPreferences(Context.MODE_PRIVATE).getBoolean(getString(R.string.pref_platform_id), true)
+
+            if (firstTime)
+                startActivity(Intent(this, ChoosePlatformActivity::class.java))
+            else
+                startActivity(Intent(this, MainActivity::class.java))
             finish()
         } else {
             showConnectionFailedDialog()
         }
-    }
-
-    private fun getPrefPlatform() {
-        val firstTime = getPreferences(Context.MODE_PRIVATE).getBoolean(getString(R.string.pref_platform_id), true)
-        if (firstTime) {
-            setPrefPlatformDialog()
-            getPreferences(Context.MODE_PRIVATE).edit().putBoolean(getString(R.string.pref_platform_id), false).apply()
-        }
-    }
-
-    private fun setPrefPlatformDialog() {
-        val builder: AlertDialog.Builder = this.let {
-            AlertDialog.Builder(it)
-        }
-        builder.apply {
-            setTitle(R.string.dialog_platform_pref_title)
-            //TODO MAAK DIT AF
-//            setItems()
-            setCancelable(false)
-        }
-        builder.create().show()
     }
 
     private fun showConnectionFailedDialog() {
@@ -115,23 +90,6 @@ class SplashScreenActivity : Activity() {
             setCancelable(false)
         }
         builder.create().show()
-    }
-
-    private fun getPlatforms() {
-        val call = getClient().getPlatforms()
-        call.enqueue(object : Callback<List<Platform>> {
-            override fun onResponse(call: Call<List<Platform>>, response: Response<List<Platform>>) {
-                platformItems = response.body()!!.toTypedArray()
-            }
-
-            override fun onFailure(call: Call<List<Platform>>, t: Throwable) {
-                Toast.makeText(
-                    this@SplashScreenActivity,
-                    getString(R.string.dialog_connection_title),
-                    Toast.LENGTH_LONG
-                ).show()
-            }
-        })
     }
 
     override fun onStop() {
